backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["daniel-jerrehian <danieljerrehian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "1.8.1"
azure-core = "1.25.0"
azure-storage-blob = "12.13.1"
cachecontrol = "0.12.11"
cachetools = "5.2.0"
certifi = "2022.6.15"
cffi = "1.15.1"
charset-normalizer = "2.1.0"
click = "8.1.3"
colorama = "0.4.5"
coverage = "6.4.4"
cryptography = "37.0.4"
deprecated = "1.2.13"
firebase-admin = "5.2.0"
flask = "2.1.3"
flask-cors = "3.0.10"
flask-marshmallow = "0.14.0"
flask-migrate = "3.1.0"
flask-sqlalchemy = "2.5.1"
gcloud = "0.18.3"
google-api-core = "2.8.2"
google-api-python-client = "2.55.0"
google-auth = "2.9.1"
google-auth-httplib2 = "0.1.0"
google-cloud-core = "2.3.2"
google-cloud-firestore = "2.6.0"
google-cloud-storage = "2.4.0"
google-crc32c = "1.3.0"
google-resumable-media = "2.3.3"
googleapis-common-protos = "1.56.4"
greenlet = "1.1.2"
grpcio = "1.47.0"
grpcio-status = "1.47.0"
httplib2 = "0.20.4"
idna = "3.3"
importlib-metadata = "4.12.0"
isodate = "0.6.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jwcrypto = "1.4.2"
jws = "0.1.3"
mako = "1.2.1"
markupsafe = "2.1.1"
marshmallow = "3.17.0"
marshmallow-sqlalchemy = "0.28.1"
msgpack = "1.0.4"
msrest = "0.7.1"
numpy = "2.0.2"
oauth2client = "4.1.3"
oauthlib = "3.2.0"
packaging = "21.3"
pandas = "2.2.2"
proto-plus = "1.20.6"
protobuf = "3.20.1"
psycopg2-binary = "2.9.3"
pyarrow = "17.0.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pycryptodome = "3.15.0"
pyparsing = "3.0.9"
pyrebase4 = "4.5.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "0.20.0"
python-jwt = "3.3.4"
pytz = "2024.2"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
requests-toolbelt = "0.9.1"
rsa = "4.9"
six = "1.16.0"
sqlalchemy = "1.4.39"
typing-extensions = "4.3.0"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "1.26.11"
werkzeug = "2.2.0"
wrapt = "1.14.1"
zipp = "3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
